// server/src/modules/setting/setting-key.enum.ts
export enum SettingKey {
  TRAINING_COOLDOWN = 'training_cooldown',
  CONTRACT_COOLDOWN = 'contract_cooldown',
  CLAN_JOIN_COOLDOWN = 'clan_join_cooldown',
  CLAN_WAR_DURATION = 'clan_war_duration',
  CLAN_WAR_COOLDOWN = 'clan_war_cooldown',
  ATTACK_COOLDOWN = 'attack_cooldown',
  PURCHASE_SHIELD_COOLDOWN = 'purchase_shield_cooldown',
  MAX_STRENGTH_FIRST_USER_GUARD = 'max_strength_first_user_guard',
  MAX_CLAN_WARS_COUNT = 'max_clan_wars_count',
  BASE_PRICE = 'base_price',
  GROWTH_FACTOR = 'growth_factor',
  MIN_INCREASE = 'min_increase',
  GUARD_BONUS = 'guard_bonus',
  POWER_PERCENTAGE = 'power_percentage',
  MAX_RANDOM_CHANCE = 'max_random_chance',
  RANDOM_POWER_COEFFICIENT = 'random_power_coefficient',
  RANDOM_BONUS_PER_GUARD = 'random_bonus_per_guard',
}